Update with logic/server sparation wit master's current changes
