+++ src/app/clusters/push-av-stream-transport-server/PushAVStreamTransportCluster.cpp
@@ -191,15 +191,6 @@ std::optional<DataModel::ActionReturnStatus> PushAvStreamTransportServer::Invoke
     return Status::UnsupportedCommand;
 }
 
-void PushAvStreamTransportServer::Shutdown()
-{
-    DefaultServerCluster::Shutdown();
-    for (const auto & timerContext : mTimerContexts)
-    {
-        DeviceLayer::SystemLayer().CancelTimer(PushAVStreamTransportDeallocateCallback, static_cast<void *>(timerContext.get()));
-    }
-}
-
 bool PushAvStreamTransportServer::HasFeature(Feature feature) const
 {
     return mFeatures.Has(feature);
@@ -358,48 +349,65 @@ void PushAvStreamTransportServer::PushAVStreamTransportDeallocateCallback(System
 
 bool PushAvStreamTransportServer::ValidateUrl(const std::string & url)
 {
-    const std::string https = "https://";
+    UriUriA uri;
+    const char * errorPos;
+    int result = uriParseSingleUriA(&uri, url.c_str(), &errorPos);
 
-    // Check minimum length and https prefix
-    if (url.size() <= https.size() || url.substr(0, https.size()) != https)
+    // Check if URI parsing failed
+    if (result != URI_SUCCESS)
     {
+        // No need to call uriFreeUriMembersA on failure
+        ChipLogError(Zcl, "PushAVStreamTransport URL : %s is not RFC 3986 compliant", url.c_str());
         return false;
     }
 
-    // Check that URL does not contain fragment character '#'
-    if (url.find('#') != std::string::npos)
+    // Extract components
+    std::string scheme = Internal::extractTextRange(uri.scheme);
+    std::string host   = Internal::extractTextRange(uri.hostText);
+    std::string path   = Internal::extractPath(uri.pathHead);
+
+    // Free URI structure
+    uriFreeUriMembersA(&uri);
+
+    // Convert scheme to lowercase for case-insensitive comparison
+    for (char & c : scheme)
     {
-        ChipLogError(Camera, "URL contains fragment character '#'");
+        c = static_cast<char>(std::tolower(static_cast<unsigned char>(c)));
+    }
+
+    if (scheme != "https")
+    {
+        ChipLogError(Camera, "URL does not have a https scheme");
         return false;
     }
 
-    // Check that URL does not contain query character '?'
-    if (url.find('?') != std::string::npos)
-    {
-        ChipLogError(Camera, "URL contains query character '?'");
-        return false;
-    }
-
-    // Check that URL ends with a forward slash '/'
-    if (url.back() != '/')
-    {
-        ChipLogError(Camera, "URL does not end with '/'");
-        return false;
-    }
-
-    // Extract host part
-    size_t hostStart = https.size();
-    size_t hostEnd   = url.find('/', hostStart);
-    std::string host = url.substr(hostStart, hostEnd - hostStart);
-
-    // Basic host validation: ensure non-empty
     if (host.empty())
     {
         ChipLogError(Camera, "URL does not contain a valid host.");
         return false;
     }
 
-    // Accept any host as long as non-empty
+    bool pathEndsWithSlash = !path.empty() && path.back() == '/';
+    if (!pathEndsWithSlash)
+    {
+        ChipLogError(Camera, "URL does not end with '/'");
+        return false;
+    }
+
+    bool noQuery = url.find('?') == std::string::npos;
+    if (!noQuery)
+    {
+        ChipLogError(Camera, "URL contains query character '?'");
+        return false;
+    }
+
+    bool noFragment = url.find('#') == std::string::npos;
+    if (!noFragment)
+    {
+        ChipLogError(Camera, "URL contains fragment character '#'");
+        return false;
+    }
+
     return true;
 }
 
@@ -441,10 +449,41 @@ Status PushAvStreamTransportServer::ValidateIncomingTransportOptions(
         transportOptions.streamUsage != StreamUsageEnum::kUnknownEnumValue, Status::ConstraintError,
         ChipLogError(Zcl, "Transport Options verification from command data[ep=%d]: Invalid streamUsage ", mPath.mEndpointId));
 
-    VerifyOrReturnValue(
-        transportOptions.videoStreamID.HasValue() || transportOptions.audioStreamID.HasValue(), Status::InvalidCommand,
-        ChipLogError(Zcl, "Transport Options verification from command data[ep=%d]: Missing videoStreamID and audioStreamID",
-                     mPath.mEndpointId));
+    // Check for video stream name length constraints
+    if (transportOptions.videoStreams.HasValue())
+    {
+        for (auto iter = transportOptions.videoStreams.Value().begin(); iter.Next();)
+        {
+            auto streamName = iter.GetValue().videoStreamName;
+            VerifyOrReturnValue(streamName.size() >= kMinStreamNameLength && streamName.size() <= kMaxStreamNameLength,
+                                Status::ConstraintError,
+                                ChipLogError(Zcl,
+                                             "Transport Options verification from command data[ep=%d]: Video Stream Name "
+                                             "Constraint Error",
+                                             mPath.mEndpointId));
+        }
+    }
+
+    // Check for audio stream name length constraints
+    if (transportOptions.audioStreams.HasValue())
+    {
+        for (auto iter = transportOptions.audioStreams.Value().begin(); iter.Next();)
+        {
+            auto streamName = iter.GetValue().audioStreamName;
+            VerifyOrReturnValue(streamName.size() >= kMinStreamNameLength && streamName.size() <= kMaxStreamNameLength,
+                                Status::ConstraintError,
+                                ChipLogError(Zcl,
+                                             "Transport Options verification from command data[ep=%d]: Audio Stream Name "
+                                             "Constraint Error",
+                                             mPath.mEndpointId));
+        }
+    }
+
+    VerifyOrReturnValue(transportOptions.TLSEndpointID <= kMaxEndpointId, Status::ConstraintError,
+                        ChipLogError(Zcl,
+                                     "Transport Options verification from command data[ep=%d]: EndpointID field Constraint Error",
+                                     mPath.mEndpointId));
+
     VerifyOrReturnValue(transportOptions.TLSEndpointID <= kMaxEndpointId, Status::ConstraintError,
                         ChipLogError(Zcl,
                                      "Transport Options verification from command data[ep=%d]: EndpointID field Constraint Error",
@@ -631,16 +670,6 @@ Status PushAvStreamTransportServer::ValidateIncomingTransportOptions(
                                          "Duration field not within allowed range",
                                          mPath.mEndpointId));
 
-        VerifyOrReturnValue(
-            containerOptions.CMAFContainerOptions.Value().trackName.size() >= 1 &&
-                containerOptions.CMAFContainerOptions.Value().trackName.size() <= kMaxTrackNameLength,
-            Status::ConstraintError,
-            ChipLogError(Zcl,
-                         "Transport Options verification from command data[ep=%d]: CMAF Container Options Track Name "
-                         "Error, actual length: %" PRIu32 " not "
-                         "within expected range of 1 to 16",
-                         mPath.mEndpointId, static_cast<uint32_t>(containerOptions.CMAFContainerOptions.Value().trackName.size())));
-
         if (containerOptions.CMAFContainerOptions.Value().CENCKey.HasValue())
         {
             VerifyOrReturnValue(
@@ -701,6 +730,223 @@ Status PushAvStreamTransportServer::ValidateIncomingTransportOptions(
     return Status::Success;
 }
 
+std::optional<DataModel::ActionReturnStatus> PushAvStreamTransportServer::ValidateStreamParameters(
+    CommandHandler & handler, const ConcreteCommandPath & commandPath,
+    const PushAvStreamTransport::Structs::TransportOptionsStruct::DecodableType & transportOptions,
+    const std::shared_ptr<PushAvStreamTransport::TransportOptionsStorage> transportOptionsPtr)
+{
+
+    // Check mutual exclusivity and presence of stream fields
+    auto listHasElements = [](const auto & list) {
+        size_t s = 0;
+        return list.ComputeSize(&s) == CHIP_NO_ERROR && s > 0;
+    };
+    bool hasNewStreamFields =
+        (transportOptions.videoStreams.HasValue() && listHasElements(transportOptions.videoStreams.Value())) ||
+        (transportOptions.audioStreams.HasValue() && listHasElements(transportOptions.audioStreams.Value()));
+    bool hasOldStreamFields = transportOptions.videoStreamID.HasValue() || transportOptions.audioStreamID.HasValue();
+
+    VerifyOrReturnValue(
+        !(hasNewStreamFields && hasOldStreamFields), Status::InvalidCommand,
+        ChipLogError(Zcl, "Transport Options[ep=%d]: Both new and old stream fields are present", mPath.mEndpointId));
+
+    VerifyOrReturnValue(hasNewStreamFields || hasOldStreamFields, Status::InvalidCommand,
+                        ChipLogError(Zcl, "Transport Options[ep=%d]: Missing stream fields", mPath.mEndpointId));
+
+    // Validate new multistream fields
+    if (hasNewStreamFields)
+    {
+        std::set<std::string> videoStreamNames;
+
+        // Validate videoStreams
+        uint16_t videoStreamsCount = 0;
+        if (transportOptions.videoStreams.HasValue())
+        {
+            std::set<uint16_t> videoStreamIDs;
+
+            for (auto iter = transportOptions.videoStreams.Value().begin(); iter.Next();)
+            {
+                auto & videoStream = iter.GetValue();
+
+                videoStreamsCount++;
+                videoStreamNames.insert(std::string(videoStream.videoStreamName.data(), videoStream.videoStreamName.size()));
+                videoStreamIDs.insert(videoStream.videoStreamID);
+            }
+
+            // Check for duplicates within video streams
+            if ((videoStreamNames.size() != videoStreamsCount) || (videoStreamIDs.size() != videoStreamsCount))
+            {
+                auto status = to_underlying(StatusCodeEnum::kDuplicateStreamValues);
+                ChipLogError(Zcl, "HandleAllocatePushTransport[ep=%d]: Duplicate videoStreamName or videoStreamID found",
+                             mPath.mEndpointId);
+                TEMPORARY_RETURN_IGNORED handler.AddClusterSpecificFailure(commandPath, status);
+                return std::nullopt;
+            }
+
+            for (auto iter = transportOptions.videoStreams.Value().begin(); iter.Next();)
+            {
+                auto & videoStream = iter.GetValue();
+                // Validate videoStreamID
+                auto status = Protocols::InteractionModel::ClusterStatusCode(mDelegate->SetVideoStream(videoStream.videoStreamID));
+                VerifyOrReturnValue(
+                    status.IsSuccess(), Status::InvalidCommand,
+                    ChipLogError(Zcl, "HandleAllocatePushTransport[ep=%d]: Invalid Video Stream ID", mPath.mEndpointId));
+            }
+        }
+
+        // Validate audioStreams
+        uint16_t audioStreamsCount = 0;
+        if (transportOptions.audioStreams.HasValue())
+        {
+            std::set<std::string> audioStreamNames;
+            std::set<uint16_t> audioStreamIDs;
+
+            // Check against video stream names
+            for (auto iter = transportOptions.audioStreams.Value().begin(); iter.Next();)
+            {
+                auto & audioStream = iter.GetValue();
+
+                audioStreamsCount++;
+                audioStreamNames.insert(std::string(audioStream.audioStreamName.data(), audioStream.audioStreamName.size()));
+                audioStreamIDs.insert(audioStream.audioStreamID);
+            }
+
+            // Check for duplicates within audio streams
+            if ((audioStreamNames.size() != audioStreamsCount) || (audioStreamIDs.size() != audioStreamsCount))
+            {
+                auto status = to_underlying(StatusCodeEnum::kDuplicateStreamValues);
+                ChipLogError(Zcl, "HandleAllocatePushTransport[ep=%d]: Duplicate audioStreamName or audioStreamID found",
+                             mPath.mEndpointId);
+                TEMPORARY_RETURN_IGNORED handler.AddClusterSpecificFailure(commandPath, status);
+                return std::nullopt;
+            }
+
+            for (auto iter = transportOptions.audioStreams.Value().begin(); iter.Next();)
+            {
+                auto & audioStream = iter.GetValue();
+                // Check for duplicate with video streams
+                if (videoStreamNames.find(std::string(audioStream.audioStreamName.data(), audioStream.audioStreamName.size())) !=
+                    videoStreamNames.end())
+                {
+                    auto status = to_underlying(StatusCodeEnum::kInvalidStream);
+                    ChipLogError(Zcl, "HandleAllocatePushTransport[ep=%d]: AudioStreamName conflicts with video stream",
+                                 mPath.mEndpointId);
+                    TEMPORARY_RETURN_IGNORED handler.AddClusterSpecificFailure(commandPath, status);
+                    return std::nullopt;
+                }
+
+                // Validate audioStreamID
+                auto status = Protocols::InteractionModel::ClusterStatusCode(mDelegate->SetAudioStream(audioStream.audioStreamID));
+                VerifyOrReturnValue(
+                    status.IsSuccess(), Status::InvalidCommand,
+                    ChipLogError(Zcl, "HandleAllocatePushTransport[ep=%d]: Invalid Audio Stream ID", mPath.mEndpointId));
+            }
+        }
+
+        // Check if both videoStreams and audioStreams are empty
+        if (videoStreamsCount == 0 && audioStreamsCount == 0)
+        {
+            auto status = to_underlying(StatusCodeEnum::kInvalidStream);
+            ChipLogError(Zcl, "Transport Options[ep=%d]: Both VideoStreams and AudioStreams empty", mPath.mEndpointId);
+            TEMPORARY_RETURN_IGNORED handler.AddClusterSpecificFailure(commandPath, status);
+            return std::nullopt;
+        }
+    }
+
+    // Validate old format single stream fields
+    if (hasOldStreamFields)
+    {
+
+        if (transportOptions.videoStreamID.HasValue())
+        {
+            uint16_t finalVideoStreamID;
+
+            if (transportOptions.videoStreamID.Value().IsNull())
+            {
+                auto delegateStatus = Protocols::InteractionModel::ClusterStatusCode(
+                    mDelegate->SelectVideoStream(transportOptions.streamUsage, finalVideoStreamID));
+                if (!delegateStatus.IsSuccess())
+                {
+                    ChipLogError(Zcl, "HandleAllocatePushTransport[ep=%d]: AllocatedVideoStreamsAttribute is empty",
+                                 mPath.mEndpointId);
+                    handler.AddStatus(commandPath, Status::InvalidInState);
+                    return std::nullopt;
+                }
+
+                transportOptionsPtr->videoStreamID.SetValue(finalVideoStreamID);
+            }
+            else
+            {
+                auto delegateStatus = Protocols::InteractionModel::ClusterStatusCode(
+                    mDelegate->SetVideoStream(transportOptions.videoStreamID.Value().Value()));
+
+                if (!delegateStatus.IsSuccess())
+                {
+                    auto cluster_status = to_underlying(StatusCodeEnum::kInvalidStream);
+                    ChipLogError(Zcl, "HandleAllocatePushTransport[ep=%d]: Invalid Video Stream ", mPath.mEndpointId);
+                    TEMPORARY_RETURN_IGNORED handler.AddClusterSpecificFailure(commandPath, cluster_status);
+                    return std::nullopt;
+                }
+
+                finalVideoStreamID = transportOptions.videoStreamID.Value().Value();
+            }
+
+            // Create new video stream entry with name 'video' and the provided or selected VideoStreamID i.e; finalVideoStreamID
+            Structs::VideoStreamStruct::Type newVideoStream;
+            newVideoStream.videoStreamName = CharSpan::fromCharString("video");
+            newVideoStream.videoStreamID   = finalVideoStreamID;
+
+            // Add to storage and update the list
+            transportOptionsPtr->AddVideoStream(newVideoStream);
+        }
+
+        if (transportOptions.audioStreamID.HasValue())
+        {
+            uint16_t finalAudioStreamID;
+
+            if (transportOptions.audioStreamID.Value().IsNull())
+            {
+                auto delegateStatus = Protocols::InteractionModel::ClusterStatusCode(
+                    mDelegate->SelectAudioStream(transportOptions.streamUsage, finalAudioStreamID));
+                if (!delegateStatus.IsSuccess())
+                {
+                    ChipLogError(Zcl, "HandleAllocatePushTransport[ep=%d]: AllocatedAudioStreamsAttribute is empty",
+                                 mPath.mEndpointId);
+                    handler.AddStatus(commandPath, Status::InvalidInState);
+                    return std::nullopt;
+                }
+
+                transportOptionsPtr->audioStreamID.SetValue(finalAudioStreamID);
+            }
+            else
+            {
+                auto delegateStatus = Protocols::InteractionModel::ClusterStatusCode(
+                    mDelegate->SetAudioStream(transportOptions.audioStreamID.Value().Value()));
+
+                if (!delegateStatus.IsSuccess())
+                {
+                    auto cluster_status = to_underlying(StatusCodeEnum::kInvalidStream);
+                    ChipLogError(Zcl, "HandleAllocatePushTransport[ep=%d]: Invalid Audio Stream ", mPath.mEndpointId);
+                    TEMPORARY_RETURN_IGNORED handler.AddClusterSpecificFailure(commandPath, cluster_status);
+                    return std::nullopt;
+                }
+
+                finalAudioStreamID = transportOptions.audioStreamID.Value().Value();
+            }
+
+            // Create new audio stream entry with name 'audio' and the provided or selected AudioStreamID i.e; finalAudioStreamID
+            Structs::AudioStreamStruct::Type newAudioStream;
+            newAudioStream.audioStreamName = CharSpan::fromCharString("audio");
+            newAudioStream.audioStreamID   = finalAudioStreamID;
+
+            // Add to storage and update the list
+            transportOptionsPtr->AddAudioStream(newAudioStream);
+        }
+    }
+
+    return Status::Success;
+}
+
 std::optional<DataModel::ActionReturnStatus>
 PushAvStreamTransportServer::HandleAllocatePushTransport(CommandHandler & handler, const ConcreteCommandPath & commandPath,
                                                          const Commands::AllocatePushTransport::DecodableType & commandData)
@@ -730,15 +976,15 @@ PushAvStreamTransportServer::HandleAllocatePushTransport(CommandHandler & handle
             commandPath.mEndpointId, handler.GetAccessingFabricIndex(), commandData.transportOptions.TLSEndpointID,
             [&](auto & TLSEndpoint) -> CHIP_ERROR {
                 // Use heap allocation for large certificate buffers to reduce stack usage
-                auto rootCertBuffer   = std::make_unique<PersistentStore<CHIP_CONFIG_TLS_PERSISTED_ROOT_CERT_BYTES>>();
-                auto clientCertBuffer = std::make_unique<PersistentStore<CHIP_CONFIG_TLS_PERSISTED_CLIENT_CERT_BYTES>>();
+                auto rootCertBuffer   = std::make_unique<PersistenceBuffer<CHIP_CONFIG_TLS_PERSISTED_ROOT_CERT_BYTES>>();
+                auto clientCertBuffer = std::make_unique<PersistenceBuffer<CHIP_CONFIG_TLS_PERSISTED_CLIENT_CERT_BYTES>>();
 
                 Tls::CertificateTable::BufferedClientCert clientCertEntry(*clientCertBuffer);
                 Tls::CertificateTable::BufferedRootCert rootCertEntry(*rootCertBuffer);
 
-                if (mTlsCertificateManagementDelegate != nullptr)
+                if (mTLSCertificateManagementDelegate != nullptr)
                 {
-                    auto & table = mTlsCertificateManagementDelegate->GetCertificateTable();
+                    auto & table = mTLSCertificateManagementDelegate->GetCertificateTable();
                     ReturnErrorOnFailure(table.GetClientCertificateEntry(handler.GetAccessingFabricIndex(),
                                                                          TLSEndpoint.ccdid.Value(), clientCertEntry));
                     ReturnErrorOnFailure(
@@ -832,7 +1078,7 @@ PushAvStreamTransportServer::HandleAllocatePushTransport(CommandHandler & handle
         TEMPORARY_RETURN_IGNORED motionZonesList.Value().Value().ComputeSize(&zoneListSize);
 
         // If there are duplicate entries, reject the command
-        std::set<uint16_t> zoneIDsFound{};
+        std::set<uint16_t> zoneIDsFound;
         bool nullFound = false;
 
         while (iterDupCheck.Next())
@@ -895,16 +1141,6 @@ PushAvStreamTransportServer::HandleAllocatePushTransport(CommandHandler & handle
         }
     }
 
-    // Validate Bandwidth Requirement
-    Status status = mDelegate->ValidateBandwidthLimit(transportOptions.streamUsage, transportOptions.videoStreamID,
-                                                      transportOptions.audioStreamID);
-    if (status != Status::Success)
-    {
-        ChipLogError(Zcl, "HandleAllocatePushTransport[ep=%d]: Resource Exhausted", mPath.mEndpointId);
-        handler.AddStatus(commandPath, status);
-        return std::nullopt;
-    }
-
     std::shared_ptr<TransportOptionsStorage> transportOptionsPtr{ new (std::nothrow) TransportOptionsStorage(transportOptions) };
 
     if (transportOptionsPtr == nullptr)
@@ -914,68 +1150,12 @@ PushAvStreamTransportServer::HandleAllocatePushTransport(CommandHandler & handle
         return std::nullopt;
     }
 
-    if (transportOptions.videoStreamID.HasValue())
+    auto validateStreamStatus = ValidateStreamParameters(handler, commandPath, transportOptions, transportOptionsPtr);
+
+    if (validateStreamStatus != Status::Success)
     {
-        if (transportOptions.videoStreamID.Value().IsNull())
-        {
-            uint16_t videoStreamID;
-
-            auto delegateStatus = Protocols::InteractionModel::ClusterStatusCode(
-                mDelegate->SelectVideoStream(transportOptions.streamUsage, videoStreamID));
-
-            if (!delegateStatus.IsSuccess())
-            {
-                handler.AddStatus(commandPath, delegateStatus);
-                return std::nullopt;
-            }
-
-            transportOptionsPtr->videoStreamID.SetValue(videoStreamID);
-        }
-        else
-        {
-            auto delegateStatus = Protocols::InteractionModel::ClusterStatusCode(
-                mDelegate->SetVideoStream(transportOptions.videoStreamID.Value().Value()));
-
-            if (!delegateStatus.IsSuccess())
-            {
-                auto cluster_status = to_underlying(StatusCodeEnum::kInvalidStream);
-                ChipLogError(Zcl, "HandleAllocatePushTransport[ep=%d]: Invalid Video Stream ", mPath.mEndpointId);
-                TEMPORARY_RETURN_IGNORED handler.AddClusterSpecificFailure(commandPath, cluster_status);
-                return std::nullopt;
-            }
-        }
-    }
-
-    if (transportOptions.audioStreamID.HasValue())
-    {
-        if (transportOptions.audioStreamID.Value().IsNull())
-        {
-            uint16_t audioStreamID;
-
-            auto delegateStatus = Protocols::InteractionModel::ClusterStatusCode(
-                mDelegate->SelectAudioStream(transportOptions.streamUsage, audioStreamID));
-
-            if (!delegateStatus.IsSuccess())
-            {
-                handler.AddStatus(commandPath, delegateStatus);
-                return std::nullopt;
-            }
-
-            transportOptionsPtr->audioStreamID.SetValue(audioStreamID);
-        }
-        else
-        {
-            auto delegateStatus = Protocols::InteractionModel::ClusterStatusCode(
-                mDelegate->SetAudioStream(transportOptions.audioStreamID.Value().Value()));
-
-            if (!delegateStatus.IsSuccess())
-            {
-                auto cluster_status = to_underlying(StatusCodeEnum::kInvalidStream);
-                ChipLogError(Zcl, "HandleAllocatePushTransport[ep=%d]: Invalid Audio Stream ", mPath.mEndpointId);
-                TEMPORARY_RETURN_IGNORED handler.AddClusterSpecificFailure(commandPath, cluster_status);
-                return std::nullopt;
-            }
-        }
+        ChipLogError(Camera, "HandleAllocatePushTransport[ep=%d]: Stream validation failed", mPath.mEndpointId);
+        return validateStreamStatus;
     }
 
     if (transportOptions.containerOptions.containerType == ContainerFormatEnum::kCmaf &&
@@ -998,6 +1178,53 @@ PushAvStreamTransportServer::HandleAllocatePushTransport(CommandHandler & handle
         }
     }
 
+    // Validate Bandwidth Requirement
+    Optional<DataModel::Nullable<uint16_t>> firstVideoStreamID;
+    Optional<DataModel::Nullable<uint16_t>> firstAudioStreamID;
+
+    if (transportOptions.videoStreams.HasValue())
+    {
+        auto videoIter = transportOptions.videoStreams.Value().begin();
+        if (videoIter.Next())
+        {
+            firstVideoStreamID.SetValue(videoIter.GetValue().videoStreamID);
+            transportOptionsPtr->videoStreamID = firstVideoStreamID;
+        }
+    }
+    if (transportOptions.audioStreams.HasValue())
+    {
+        auto audioIter = transportOptions.audioStreams.Value().begin();
+        if (audioIter.Next())
+        {
+            firstAudioStreamID.SetValue(audioIter.GetValue().audioStreamID);
+            transportOptionsPtr->audioStreamID = firstAudioStreamID;
+        }
+    }
+
+    Status retStatus = mDelegate->ValidateBandwidthLimit(transportOptions.streamUsage, firstVideoStreamID, firstAudioStreamID);
+
+    if (retStatus != Status::Success)
+    {
+        ChipLogError(Zcl, "HandleAllocatePushTransport[ep=%d]: Resource Exhausted", mPath.mEndpointId);
+        handler.AddStatus(commandPath, retStatus);
+        return std::nullopt;
+    }
+
+    // Validate MaxPreRollLength constraint
+    if (transportOptionsPtr->videoStreamID.HasValue() && transportOptions.triggerOptions.maxPreRollLen.HasValue())
+    {
+        uint16_t maxPreRollLength = transportOptions.triggerOptions.maxPreRollLen.Value();
+        if (maxPreRollLength != 0 &&
+            !mDelegate->ValidateMaxPreRollLength(maxPreRollLength, transportOptionsPtr->videoStreamID.Value()))
+        {
+            auto maxPreRollLengthStatus = to_underlying(StatusCodeEnum::kInvalidPreRollLength);
+            ChipLogError(Zcl, "HandleAllocatePushTransport[ep=%d]: MaxPreRollLength (%u) validation failed", mPath.mEndpointId,
+                         maxPreRollLength);
+            TEMPORARY_RETURN_IGNORED handler.AddClusterSpecificFailure(commandPath, maxPreRollLengthStatus);
+            return std::nullopt;
+        }
+    }
+
     uint16_t connectionID = GenerateConnectionID();
 
     if (connectionID == kMaxConnectionId)
@@ -1009,9 +1236,9 @@ PushAvStreamTransportServer::HandleAllocatePushTransport(CommandHandler & handle
 
     FabricIndex accessingFabricIndex = handler.GetAccessingFabricIndex();
 
-    status = mDelegate->AllocatePushTransport(*transportOptionsPtr, connectionID, accessingFabricIndex);
+    retStatus = mDelegate->AllocatePushTransport(*transportOptionsPtr, connectionID, accessingFabricIndex);
 
-    if (status == Status::Success)
+    if (retStatus == Status::Success)
     {
         // add connection to CurrentConnections
         TransportConfigurationStorage outTransportConfiguration(connectionID, transportOptionsPtr);
@@ -1034,7 +1261,7 @@ PushAvStreamTransportServer::HandleAllocatePushTransport(CommandHandler & handle
     }
     else
     {
-        handler.AddStatus(commandPath, status);
+        handler.AddStatus(commandPath, retStatus);
     }
 
     return std::nullopt;
@@ -1130,6 +1357,14 @@ PushAvStreamTransportServer::HandleModifyPushTransport(CommandHandler & handler,
         ChipLogError(Zcl, "HandleModifyPushTransport[ep=%d]: Memory Allocation failed for transportOptions", mPath.mEndpointId);
         handler.AddStatus(commandPath, Status::ResourceExhausted);
         return std::nullopt;
+    }
+
+    auto validateStreamStatus = ValidateStreamParameters(handler, commandPath, transportOptions, transportOptionsPtr);
+
+    if (validateStreamStatus != Status::Success)
+    {
+        ChipLogError(Camera, "HandleAllocatePushTransport[ep=%d]: Stream validation failed", mPath.mEndpointId);
+        return validateStreamStatus;
     }
 
     // Call the delegate
@@ -1314,7 +1549,29 @@ PushAvStreamTransportServer::HandleManuallyTriggerTransport(CommandHandler & han
 
     if (status == Status::Success)
     {
-        GeneratePushTransportBeginEvent(connectionID, TransportTriggerTypeEnum::kCommand, MakeOptional(activationReason));
+        // Get container type from transport configuration
+        ContainerFormatEnum containerType = transportConfiguration->transportOptions.Value().containerOptions.containerType;
+
+        // For CMAF container type, we need to provide CMAF session number
+        Optional<uint64_t> cmafSessionNumber;
+        if (containerType == ContainerFormatEnum::kCmaf)
+        {
+            uint64_t sessionNumber = 0;
+            if (mDelegate != nullptr && mDelegate->GetCMAFSessionNumber(connectionID, sessionNumber))
+            {
+                cmafSessionNumber = MakeOptional<uint64_t>(sessionNumber);
+            }
+            else
+            {
+                ChipLogError(Zcl, "GeneratePushTransportBeginEvent: Unable to get CMAF session number for connection %u",
+                             connectionID);
+                // Don't include the session number if we can't get it
+                cmafSessionNumber = Optional<uint64_t>();
+            }
+        }
+
+        GeneratePushTransportBeginEvent(connectionID, TransportTriggerTypeEnum::kCommand, MakeOptional(activationReason),
+                                        containerType, cmafSessionNumber);
     }
 
     handler.AddStatus(commandPath, status);
@@ -1442,14 +1699,18 @@ Status PushAvStreamTransportServer::CheckPrivacyModes(StreamUsageEnum streamUsag
 
 Status PushAvStreamTransportServer::GeneratePushTransportBeginEvent(const uint16_t connectionID,
                                                                     const TransportTriggerTypeEnum triggerType,
-                                                                    const Optional<TriggerActivationReasonEnum> activationReason)
+                                                                    const Optional<TriggerActivationReasonEnum> activationReason,
+                                                                    const ContainerFormatEnum containerType,
+                                                                    const Optional<uint64_t> cmafSessionNumber)
 {
     Events::PushTransportBegin::Type event;
     EventNumber eventNumber;
 
-    event.connectionID     = connectionID;
-    event.triggerType      = triggerType;
-    event.activationReason = activationReason;
+    event.connectionID      = connectionID;
+    event.triggerType       = triggerType;
+    event.activationReason  = activationReason;
+    event.containerType     = containerType;
+    event.CMAFSessionNumber = cmafSessionNumber;
 
     CHIP_ERROR err = LogEvent(event, mPath.mEndpointId, eventNumber);
     if (CHIP_NO_ERROR != err)
@@ -1467,6 +1728,40 @@ Status PushAvStreamTransportServer::GeneratePushTransportEndEvent(const uint16_t
     EventNumber eventNumber;
 
     event.connectionID = connectionID;
+
+    // Get container type from transport configuration to include in the event
+    TransportConfigurationStorage * transportConfig = FindStreamTransportConnection(connectionID);
+    if (transportConfig != nullptr)
+    {
+        event.containerType = transportConfig->transportOptions.Value().containerOptions.containerType;
+
+        // For CMAF container type, we need to provide CMAF session number
+        if (event.containerType == ContainerFormatEnum::kCmaf)
+        {
+            uint64_t sessionNumber = 0;
+            if (mDelegate != nullptr && mDelegate->GetCMAFSessionNumber(connectionID, sessionNumber))
+            {
+                event.CMAFSessionNumber = MakeOptional<uint64_t>(sessionNumber);
+            }
+            else
+            {
+                ChipLogError(Zcl, "GeneratePushTransportEndEvent: Unable to get CMAF session number for connection %u",
+                             connectionID);
+                // Don't include the session number if we can't get it
+                event.CMAFSessionNumber = Optional<uint64_t>();
+            }
+        }
+        else
+        {
+            event.CMAFSessionNumber = Optional<uint64_t>();
+        }
+    }
+    else
+    {
+        // Fallback values if transport config not found
+        event.containerType     = ContainerFormatEnum::kCmaf; // Default fallback
+        event.CMAFSessionNumber = Optional<uint64_t>();
+    }
 
     CHIP_ERROR err = LogEvent(event, mPath.mEndpointId, eventNumber);
     if (CHIP_NO_ERROR != err)
@@ -1492,8 +1787,37 @@ Status PushAvStreamTransportServer::NotifyTransportStarted(uint16_t connectionID
         return Status::NotFound;
     }
 
+    // Get container type from transport configuration
+    ContainerFormatEnum containerType = transportConfig->transportOptions.Value().containerOptions.containerType;
+
+    // For CMAF container type, we need to provide CMAF session number
+    Optional<uint64_t> cmafSessionNumber;
+    if (containerType == ContainerFormatEnum::kCmaf)
+    {
+        uint64_t sessionNumber = 0;
+        if (mDelegate != nullptr && mDelegate->GetCMAFSessionNumber(connectionID, sessionNumber))
+        {
+            cmafSessionNumber = MakeOptional<uint64_t>(sessionNumber);
+        }
+        else
+        {
+            ChipLogError(Zcl, "NotifyTransportStarted: Unable to get CMAF session number for connection %u", connectionID);
+            // Don't include the session number if we can't get it
+            cmafSessionNumber = Optional<uint64_t>();
+        }
+    }
+
     // Generate the PushTransportBegin event
-    return GeneratePushTransportBeginEvent(connectionID, triggerType, activationReason);
+    return GeneratePushTransportBeginEvent(connectionID, triggerType, activationReason, containerType, cmafSessionNumber);
+}
+
+void PushAvStreamTransportServer::Shutdown(ClusterShutdownType shutdownType)
+{
+    DefaultServerCluster::Shutdown(shutdownType);
+    for (const auto & timerContext : mTimerContexts)
+    {
+        DeviceLayer::SystemLayer().CancelTimer(PushAVStreamTransportDeallocateCallback, static_cast<void *>(timerContext.get()));
+    }
 }
 
 Status PushAvStreamTransportServer::NotifyTransportStopped(uint16_t connectionID, TransportTriggerTypeEnum triggerType)
@@ -1513,6 +1837,37 @@ Status PushAvStreamTransportServer::NotifyTransportStopped(uint16_t connectionID
     return GeneratePushTransportEndEvent(connectionID);
 }
 
+namespace Internal {
+
+std::string extractTextRange(const UriTextRangeA & range)
+{
+    if (range.first == nullptr || range.afterLast == nullptr || range.first >= range.afterLast)
+    {
+        return "";
+    }
+    return std::string(range.first, range.afterLast);
+}
+
+std::string extractPath(const UriPathSegmentA * pathHead)
+{
+    std::string path;
+    const UriPathSegmentA * segment = pathHead;
+
+    while (segment != nullptr)
+    {
+        path += "/";
+
+        if (segment->text.first != nullptr && segment->text.afterLast != nullptr)
+        {
+            path += std::string(segment->text.first, segment->text.afterLast);
+        }
+        segment = segment->next;
+    }
+
+    return path;
+}
+
+} // namespace Internal
 } // namespace Clusters
 } // namespace app
 } // namespace chip
